refactor(planRecommendations): refactored api call and request

+++ services/planning/src/main/java/com/petek/planning/dto/ForecastRequest.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ForecastRequest {
-    private List<Integer> productIds;
+    private List<Long> productIds;
     private Integer horizonDays;
     private String asOfDate;
     private Boolean returnDaily;
