+++ services/planning/src/main/java/com/petek/planning/service/WebClientService.java
@@ -36,7 +36,7 @@ public class WebClientService {
         );
     }
 
-    public String getForecasts(Integer productId, int horizonDays, LocalDate asOfDate) {
+    public String getForecasts(Long productId, int horizonDays, LocalDate asOfDate) {
         ForecastRequest request = ForecastRequest.builder()
             .productIds(List.of(productId))
             .horizonDays(horizonDays)
@@ -58,7 +58,7 @@ public class WebClientService {
             entity,
             String.class
         );
-
+        
         return response.getBody();
     }
